+++ udata_hydra/utils/http.py
@@ -48,4 +48,11 @@ async def send(dataset_id: str, resource_id: str, document: dict) -> None:
     async with aiohttp.ClientSession() as session:
         async with session.put(uri, json=document, headers=headers) as resp:
             # we're raising since we should be in a worker thread
-            resp.raise_for_status()
+            if resp.status == 404:
+                pass
+            elif resp.status == 410:
+                raise IOError("Resource has been deleted on udata")
+            if resp.status == 502:
+                raise IOError("Udata is unreachable")
+            else:
+                resp.raise_for_status()