+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 - Fix Sentry issue #4195 [#209](https://github.com/datagouv/hydra/pull/209)
 - Clean doctrings for more consistent style [#215](https://github.com/datagouv/hydra/pull/215)
 - Fix some type hints [#214](https://github.com/datagouv/hydra/pull/214)
-- Add option to force analysis even if resource has not changed[#205](https://github.com/datagouv/hydra/pull/205)
+- Add option to force analysis even if resource has not changed [#205](https://github.com/datagouv/hydra/pull/205)
 - Fix get all checks CRUD method [#217](https://github.com/datagouv/hydra/pull/217)
 - Deactivate parquet export for small CSVs [#216](https://github.com/datagouv/hydra/pull/216)