+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Refactor ParseError to enrich Sentry with context and to inquire about Sentry errors like #4096  [#218](https://github.com/datagouv/hydra/pull/218)
 - Remove legacy routes [#203](https://github.com/datagouv/hydra/pull/203)
+- More explicit error reporting when sending to udata without raising errors for udata responding with a 404 [#213](https://github.com/datagouv/hydra/pull/213)
 
 ## 2.0.5 (2024-11-08)
 
@@ -13,7 +14,7 @@
 - Fix Sentry issue #4195 [#209](https://github.com/datagouv/hydra/pull/209)
 - Clean doctrings for more consistent style [#215](https://github.com/datagouv/hydra/pull/215)
 - Fix some type hints [#214](https://github.com/datagouv/hydra/pull/214)
-- Add option to force analysis even if resource has not changed[#205](https://github.com/datagouv/hydra/pull/205)
+- Add option to force analysis even if resource has not changed [#205](https://github.com/datagouv/hydra/pull/205)
 - Fix get all checks CRUD method [#217](https://github.com/datagouv/hydra/pull/217)
 - Deactivate parquet export for small CSVs [#216](https://github.com/datagouv/hydra/pull/216)